Adsorption

+++ src/Ewald.cpp
@@ -21,7 +21,6 @@ along with this program, also can be found at <http://www.gnu.org/licenses/>.
 #include "GeomLib.h"
 #include "NumLib.h"
 #include <cassert>
-#include <omp.h>
 #ifdef GOMC_CUDA
 #include "CalculateEwaldCUDAKernel.cuh"
 #include "ConstantDefinitionsCUDAKernel.cuh"
@@ -141,7 +140,7 @@ void Ewald::RecipInit(uint box, BoxDimensions const& boxAxes)
 
    for (x = 0; x <= kmax[box]; x++)
    {
-      nky_max = sqrt(pow(kmax[box], 2) - pow(x, 2));
+      nky_max = sqrt(pow((double)kmax[box], 2) - pow((double)x, 2));
       nky_min = -nky_max;
       if (x == 0.0)
       {
@@ -149,7 +148,7 @@ void Ewald::RecipInit(uint box, BoxDimensions const& boxAxes)
       }
       for (y = nky_min; y <= nky_max; y++)
       {
-	 nkz_max = sqrt(pow(kmax[box], 2) - pow(x, 2) - pow(y, 2));
+	 nkz_max = sqrt(pow((double)kmax[box], 2) - pow((double)x, 2) - pow((double)y, 2));
 	 nkz_min = -nkz_max;
 	 if (x == 0.0 && y == 0.0)
          {
